moving run-async directive into its own directory
